LPFG-569: Created MediaEntity abstract class
