+++ src/main/java/uk/gov/cslearning/catalogue/dto/factory/MediaFactory.java
@@ -3,7 +3,7 @@ package uk.gov.cslearning.catalogue.dto.factory;
 import com.google.common.collect.ImmutableMap;
 import org.springframework.stereotype.Component;
 import uk.gov.cslearning.catalogue.domain.media.Document;
-import uk.gov.cslearning.catalogue.domain.media.Media;
+import uk.gov.cslearning.catalogue.domain.media.MediaEntity;
 import uk.gov.cslearning.catalogue.dto.FileUpload;
 
 import java.time.LocalDateTime;
@@ -13,7 +13,7 @@ import java.util.function.Function;
 @Component
 public class MediaFactory {
 
-    private Map<String, Function<FileUpload, Media>> createMethods = ImmutableMap.of(
+    private Map<String, Function<FileUpload, MediaEntity>> createMethods = ImmutableMap.of(
             "doc", fileUpload -> {
                 Document document = new Document();
                 document.setContainer(fileUpload.getContainer());
@@ -30,7 +30,7 @@ public class MediaFactory {
 
 
 
-    public Media create(FileUpload fileUpload) {
+    public MediaEntity create(FileUpload fileUpload) {
         return createMethods.get(fileUpload.getExtension()).apply(fileUpload);
     }
 }
